requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
websockets
google-generativeai
python-dotenv
chromadb
pydantic